+++ NEWS.md
@@ -1,3 +1,5 @@
+# ami (development version)
+
 # ami 0.2.0
 
 * Added `using_quarto()` for detecting whether code is running as part of a Quarto document (#25)