+++ NEWS.md
@@ -1,6 +1,7 @@
 # ami (development version)
 
-* Added `using_positron()` and helpers `using_positron_desktop()` and `using_positron_server()` to detect the Positron IDE.
+* Added `using_positron()` and helpers `using_positron_desktop()` and `using_positron_server()` to detect the Positron IDE (#29)
+* Added `using_python_venv()` to detect Python virtual environments
 
 
 # ami 0.2.0